+++ src/model/Model_StockHistory.cpp
@@ -70,7 +70,7 @@ DB_Table_STOCKHISTORY_V1::DATE Model_StockHistory::DATE(const wxDate& date, OP o
 /**
 Adds or updates an element in stock history
 */
-int Model_StockHistory::addUpdate(const wxString& symbol, const wxDate& date, double price, UPDTYPE type)
+int64 Model_StockHistory::addUpdate(const wxString& symbol, const wxDate& date, double price, UPDTYPE type)
 {
     Data *stockHist = this->get(symbol, date);
     if (!stockHist) stockHist = this->create();
